updated db migration folder structure

+++ src/Reliance.Web.Client/Api/ProjectDto.cs
@@ -2,6 +2,6 @@
 {
     public class ProjectDto: BaseDto
     {
-        public int SolutionId { get; set; }
+        public long SolutionId { get; set; }
     }
 }
