+++ src/Reliance.Web.Client/Api/RepositoryOwnerApiKeyDto.cs
@@ -6,8 +6,8 @@ namespace Reliance.Web.Client.Api
 {
     public class RepositoryOwnerApiKeyDto
     {
-        public int Id { get; set; }
-        public int OwnerId { get; set; }
+        public long Id { get; set; }
+        public long OwnerId { get; set; }
         public string ApiKey { get; set; }
         public DateTime ExpireOn { get; set; }
         public bool IsRevoked { get; set; }
